Warn when a namespace type is not supported
Instead of aborting execution, write instead a warning message to the standard
error stream, trying to continue execution anyway. Most of the time when a file
under /proc/<pid>/ns/ is missing, it means that kernel support is not available.
For example, support for user namespaces is often disabled in kernels supplied
by distributions, but certain kinds of container run anyway, so "doin" should
try and work anyway in those cases.